tests/wpt/web-platform-tests/websockets/Create-Secure-blocked-port.htm

<!DOCTYPE html>
<html>
<head>
<title>W3C WebSocket API - Create Secure WebSocket - blocked port</title>
<script type="text/javascript" src="/resources/testharness.js"></script>
<script type="text/javascript" src="/resources/testharnessreport.js"></script>
<script type="text/javascript" src="websocket.js?pipe=sub"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function () {
  assert_true("WebSocket" in window);
  var blockedPort = 25;
  var url = "wss://" + __SERVER__NAME + ":" + blockedPort + "/" + __PATH;
  assert_throws("SECURITY_ERR", function() {
    new WebSocket(url)
  });
}, "W3C WebSocket API - Create Secure WebSocket - Pass a URL with a blocked port - SECURITY_ERR should be thrown")
</script>
</body>
</html>
